proxy & redirect to auth/user route
