+++ infra/k8s/ingress-srv.yaml
@@ -1,0 +1,18 @@
+apiVersion: networking.k8s.io/v1
+kind: Ingress
+metadata:
+  name: api-conexa
+  annotations:
+    kubernetes.io/ingress.class: nginx
+    nginx.ingress.kubernetes.io/use-regex: "true"
+spec:
+  rules:
+  -  http:
+      paths:
+        - path: /.*
+          pathType: Prefix
+          backend:
+            service:
+              name: login-srv
+              port:
+                number: 3000
