+++ infra/k8s/login-srv.yaml
@@ -1,0 +1,48 @@
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: login-depl
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: login
+  template:
+    metadata:
+      labels:
+        app: login
+    spec:
+      containers:
+        - name: login
+          image: login
+          imagePullPolicy: Never
+          env:
+          - name: APP_PORT
+            value: "3000"
+          - name: MONGO_HOST
+            value: "login-mongo-srv"
+          - name: MONGO_PORT
+            value: "27017"
+          - name: NATS_HOST
+            value: "nats-srv"
+          - name: NATS_PORT
+            value: "4222"           
+          - name: REDIS_HOST
+            value: "login-redis-srv"
+          - name: REDIS_PORT
+            value: "6379"
+
+
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: login-srv
+spec:
+  selector:
+    app: login
+  ports:
+    - name: login
+      protocol: TCP
+      port: 3000
+      targetPort: 3000
