+++ login/lib/entities/authProxy/authProxy.routes.js
@@ -4,8 +4,8 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 const express_1 = __importDefault(require("express"));
+const authProxy_controller_1 = require("../authProxy/authProxy.controller");
 const router = express_1.default.Router();
-router.get('/', (req, res) => {
-    res.json({ msg: 'proxy' });
-});
+router.get('/', authProxy_controller_1.getAll);
+router.get('/user/:email', authProxy_controller_1.search);
 exports.default = router;
