+++ login/src/entities/user/user.controller.ts
@@ -37,25 +37,3 @@ export const login = async (req: Request, res:Response) => {
     return res.status(400).json({msg: 'Invalid email or password'})
 }
 
-export const getAll = async (req: Request, res:Response) => {
-    res.status(400).json({msg: 'proxy'})
-    // res.status(200).json({msg: 'proxy'})
-    // const user = await User.findOne({email: req.body.email})
-    // const verifyToken = token.verifyToken(req.body.token)
-
-    // !user ? res.send(400).json({msg: 'User is not registered'}) : verifyToken
-    // !verifyToken ? res.send(401).json({msg: 'Token expired, please login again'}) : 
-    // res.status(200).json(await User.find().limit(10))
-}
-
-export const search = async (req: Request, res:Response) => {
-    const user = await User.findOne({email: req.body.email})
-    const verifyToken = token.verifyToken(req.body.token)
-
-    !user ? res.send(400).json({msg: 'User is not registered'}) : verifyToken
-    !verifyToken ? res.send(401).json({msg: 'Token expired, please login again'}) : 
-    res.status(200).json(await User.find({email: {
-        $regex: req.params.email,
-        $options: "i"
-      }} ))
-}
